Develop

Closes S03P13B103-109, S03P13B103-110, S03P13B103-88, S03P13B103-94, S03P13B103-113, S03P13B103-114, S03P13B103-68, S03P13B103-64, S03P13B103-60, S03P13B103-107, S03P13B103-108, S03P13B103-89, S03P13B103-96, and S03P13B103-06

See merge request s03-webmobile1-sub3/s03p13b103!20

+++ Frontend/src/views/user/userPage.vue
@@ -3,7 +3,19 @@
       <v-row justify="center" id="userpage" class="mx-10 my-10">
           <v-col cols="7" class="flex-column">
             <v-row class="d-flex align-center mr-2">
-                <img :src="$store.state.baseUrl+`/user`+userInfo.userImg" id="userImg" class="mr-10 ml-5 my-5" />
+                <v-col class="d-flex flex-column justify-center">
+                    <img :src="imgSrc" ref="imgpreview" id="userImg" class="mr-10 ml-5 my-5" />
+                    <v-row class="d-flex flex-nowrap justify-center">
+                        <v-file-input
+                            v-model="userInfo.userImg"
+                            accept="image/*" 
+                            @change="changeImg"
+                        ></v-file-input>
+                        <v-btn depressed rounded small class="mt-5 mx-1" @click="resetImg">
+                            초기화
+                        </v-btn>
+                    </v-row>
+                </v-col>
                 <v-col class="d-flex flex-column justify-center">
                     <v-row class="mb-5">
                         <h1 style="color:black">{{userInfo.userEmail}}</h1>
@@ -103,6 +115,7 @@
 
 <script>
 import axios from 'axios';
+import querystring from 'querystring';
 
 export default {
     name: "userPage",
@@ -112,23 +125,21 @@ export default {
     
     data() {
         return {
-            userInfo: {},
-            newUserInfo: {
-                userEmail: "",
-                userPw: "",
-                userNm: "",
-                userImg: "",
-                userIntro: "",
-                userGoalHr: ""
-            },
             updatePwdDialog: false,
             deleteUserDialog: false,
+            userInfo: {},
+            imgInfo: {
+                originSrc: this.$store.state.comm.baseUrl+"/image/user",
+                default: this.$store.state.comm.baseUrl+"/image/user/userDefault.png",
+                updateImg: false,
+                correctExt: true,
+            },
             updatePwd: {
                 origin: "",
                 new: "",
                 newChk: "",
             },
-
+            formData: null
         }
     },
     
@@ -138,32 +149,60 @@ export default {
     
     methods: {
         getUserInfo() {
-            axios.get('/user/')
+            axios.get('/user')
             .then(({ data }) => {
                 console.log(data);
                 this.userInfo = data.object;
-                this.newUserInfo.userPw = "";
-                this.newUserInfo.userNm = this.userInfo.userNm;
-                this.newUserInfo.userImg = this.userInfo.userImg;
-                this.newUserInfo.userIntro = this.userInfo.userIntro;
-                this.newUserInfo.userGoalHr = this.userInfo.userGoalHr;
+                this.imgInfo.originSrc = this.imgInfo.originSrc.concat(this.userInfo.userImg);
+                this.imgSrc =this.imgInfo.originSrc;
+                this.userInfo.userImg = "";
+                console.log(this.imgInfo.originSrc);
             })
-            
+        },
+
+        changeImg(e){
+            console.log(e);
+            // const file = e.target.files[0]; // Get first index in files
+            this.imgInfo.correctExt = false;
+            if (e) {
+                this.confirmExt();
+                if (this.imgInfo.correctExt) {
+                    this.$refs.imgpreview.src = e ? URL.createObjectURL(e) : this.imgInfo.originSrc;
+                    this.imgInfo.updateImg = true;
+                } else {
+                    alert("지원하지 않는 확장자입니다.");
+                    this.userInfo.userImg = "";
+                }
+            }
+        },
+
+        resetImg(){
+            this.imgInfo.updateImg = true;
+            this.$refs.imgpreview.src = this.imgInfo.default;
+            this.userInfo.userImg = "";
+        },
+
+        confirmExt() {
+            this.imgInfo.correctExt = false;
+            const ext = this.userInfo.userImg.name.substring(this.userInfo.userImg.name.lastIndexOf(".")+1, this.userInfo.userImg.name.length).toLowerCase();
+            const imgExts = "xbm,tif,pjp,pjpeg,svgz,jpg,jpeg,ico,tiff,gif,svg,bmp,png,gfif,webp";
+            const eachExts = imgExts.split(",");
+
+            for (let i = 0; i < eachExts.length; i++) {
+                if (ext == eachExts[i]) this.imgInfo.correctExt = true;
+            }
         },
 
         checkUserPwd(pwd, type) {
-            console.log(pwd); 
-            axios.post('/user/checkPw', {
-                    userPw: pwd
-            })
+            console.log("왓니?");
+            axios.post('/user/checkpw', querystring.stringify({userPw : pwd}))
             .then(({ data }) => {
                 console.log(data);
+                console.log("????");
                 if (data.msg === "success" && data.object) {
                     if (type === "update") {
-                        console.log("업데이트ㄱㄱ");
                         this.updateUserPwd();
                     } else if (type === "del") {
-                        console.log("삭제ㄱㄱ");
                         this.deleteUser();
                     }
                 } else if (data.msg === "success" && !data.object) {
@@ -171,29 +210,70 @@ export default {
                 } else {
                     alert("오류 발생");
                 }
+            }).catch(({ err }) => {
+                console.log(err);
             })
+        },
+        
+        makeFormData() {
+            this.formData = new FormData();
+
+            this.formData.append('userNm', this.userInfo.userNm);
+            if (this.userInfo.userImg instanceof File) {
+                this.formData.append('userImg', this.userInfo.userImg);
+            }
+            this.formData.append('userIntro', this.userInfo.userIntro);
+            this.formData.append('userGoalHr', Number(this.userInfo.userGoalHr));
+            this.formData.append('updateImg', this.imgInfo.updateImg);
         },
 
         updateUserInfo() {
-            axios.put('/user', {
+            this.makeFormData();
+
+            const config = {
+                headers: {
+                'Content-Type' : 'multipart/form-data',
+                }
+            }
+            
+            for(const pair of this.formData.entries()) {
+                console.log(pair[0]+ ', '+ pair[1]); 
+            }
+
+            axios.put('/user', this.formData, config)
+            .then(({ data }) => {
+                console.log(data);
+                if (data.msg === "success") {
+                    alert("회원 정보 수정 완료");
+                    this.$router.go();
+                }
             })
 
         },
 
         updateUserPwd() {
-            this.newUserInfo.userPw = this.updatePwd.new;
-            axios.get('/user', this.newUserInfo)
+            axios.put('/user', querystring.stringify({userNewPw : this.updatePwd.new}))
             .then(({ data }) => {
                 console.log(data);
+                if (data.msg === "success") {
+                    alert("비밀번호 변경 성공");
+                    this.updatePwdDialog=false;
+                }
             })
         },
 
         deleteUser() {
-            axios.delete('/user/'+this.userInfo.userId)
-
+            axios.delete('/user')
+            .then(({ data }) => {
+                console.log(data);
+                if (data.msg === "success") {
+                    alert("탈퇴가 완료되었습니다.")
+                    this.$store.commit("logoutSuccess");
+                    this.$router.push({ name: "Home" });
+                }
+            })
         },
     }
-
 }
 </script>
